Refactor bot scanning and navigation logic
- Updated BotSceneScan3D to improve block and entity scanning with new BotScanInfo class.
- Changed scan radius and height parameters from double to int for consistency.
- Simplified block data handling by using toBlockData() method.
- Enhanced distance calculations in BlockUtils for better performance.
- Refactored BotPosition and BotBlockData to provide clearer access to coordinates.
- Improved comparator implementations for BotBlockData and BotPosition.
- Added new methods for yaw/pitch calculations in BotPositionSight for better targeting.
- Removed deprecated methods and cleaned up code for readability and maintainability.

+++ src/main/java/com/devone/bot/core/task/active/sonar/BotSonar3DTask.java
@@ -10,7 +10,7 @@ import com.devone.bot.core.utils.logger.BotLogger;
 
 public class BotSonar3DTask extends BotTaskAutoParams<BotSonarTaskParams> {
 
-    private double radius;
+    private int radius;
     private int height;
 
     public BotSonar3DTask(Bot bot) {
