+++ src/main/java/com/devone/bot/core/utils/BotConstants.java
@@ -17,7 +17,7 @@ public class BotConstants {
 
     public static final boolean FLIP_COORDS = false;
 
-    public static final double DEFAULT_SCAN_RANGE = 5.0;
+    public static final int DEFAULT_SCAN_RANGE = 5;
     public static final int DEFAULT_SCAN_DATA_SLICE_HEIGHT = 4;
     public static final double DEFAULT_NORMAL_SIGHT_FOV = 45.0;
     public static final double DEFAULT_MAX_SIGHT_FOV = 360.0;
